1.去除了重复的create函数，将获取字典的步骤移到了ConfRoot中。 2.补充测试。 3.更新README

+++ conf_root/agents/JsonAgent.py
@@ -11,17 +11,6 @@ class JsonAgent(BasicAgent):
     def exist(self, configuration: Configuration) -> bool:
         return self.get_configuration_location(configuration).exists()
 
-    def create(self, configuration: Configuration):
-        super().create(configuration)
-        location = self.get_configuration_location(configuration)
-        # 将默认值转换为dict，便于序列化
-        default_dict = configuration.defaults
-
-        # 将dict转换为YAML并写入文件
-        if len(default_dict.keys()) > 0:
-            with open(location, "w") as file:
-                json.dump(default_dict, file)
-
     def load(self, configuration):
         super().load(configuration)
         location = self.get_configuration_location(configuration)
@@ -30,10 +19,12 @@ class JsonAgent(BasicAgent):
         # 将dict展开为对象。
         return data
 
-    def save(self, configuration, obj):
-        super().save(configuration, obj)
+    def save(self, configuration, data):
+        super().save(configuration, data)
         location = self.get_configuration_location(configuration)
-        data_dict = configuration.obj2data(obj)
+        if not os.path.exists(location) and len(data.keys()) == 0:
+            return
+
         # 将dict转换为YAML并写入文件
         with open(location, "w") as file:
-            json.dump(data_dict, file)
+            json.dump(data, file)
