+++ conf_root/agents/YamlAgent.py
@@ -12,21 +12,11 @@ class YamlAgent(BasicAgent):
     def __init__(self, location):
         super().__init__(location)
         self.yaml = YAML()
+        self.yaml.preserve_quotes = True
 
     def exist(self, configuration: Configuration) -> bool:
         location = self.get_configuration_location(configuration)
         return os.path.exists(location)
-
-    def create(self, configuration: Configuration):
-        super().create(configuration)
-        location = self.get_configuration_location(configuration)
-        # 将dataclass默认值转换为dict，便于序列化
-        default_dict = configuration.defaults
-
-        # 将dict转换为YAML并写入文件
-        if len(default_dict.keys()) > 0:
-            with open(location, "w") as file:
-                self.yaml.dump(default_dict, file)
 
     def load(self, configuration) -> Dict[str, Any]:
         super().load(configuration)
@@ -36,10 +26,11 @@ class YamlAgent(BasicAgent):
         # 将dict展开为对象。
         return data
 
-    def save(self, configuration, obj):
-        super().save(configuration, obj)
+    def save(self, configuration, data):
+        super().save(configuration, data)
         location = self.get_configuration_location(configuration)
-        data_dict = configuration.obj2data(obj)
+        if not os.path.exists(location) and len(data.keys()) == 0:
+            return
         # 将dict转换为YAML并写入文件
         with open(location, "w") as file:
-            self.yaml.dump(data_dict, file)
+            self.yaml.dump(data, file)
